requirements.txt
flask==2.3.3
openai==1.54.4
python-dotenv==1.0.0
markupsafe==2.1.3
markdown==3.5.1
